# RuGram/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.MetaData:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      totalPages:
        type: integer
    type: object
  dto.PaginationResponse:
    properties:
      data: {}
      meta:
        $ref: '#/definitions/dto.MetaData'
    type: object
...
  dto.UpdateUserRequest:
    properties:
      email:
        type: string
      password:
        type: string
      phone:
        type: string
    type: object
  dto.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      phone:
        type: string
      updated_at:
        type: string
    type: object
...
  utils.ErrorResponseStruct:
    properties:
      error:
        type: string
      success:
        type: boolean
    type: object
  utils.StandardResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:4200
info:
  contact:
    email: support@rugram.com
    name: Gleb, Pavel
  description: API for RuGram social platform with MongoDB, JWT, OAuth, Redis.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: RuGram API
  version: "1.0"
paths:
...
  /posts/user/{userId}:
    get:
      consumes:
      - application/json
      description: Returns paginated posts of a specific user
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        maximum: 100
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.PaginationResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponseStruct'
      security:
      - BearerAuth: []
      summary: Get posts by user ID
      tags:
      - Posts
  /users:
    get:
      consumes:
      - application/json
      description: Returns a paginated list of all users
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        maximum: 100
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.PaginationResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponseStruct'
      security:
      - BearerAuth: []
      summary: List users with pagination
      tags:
      - Users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Soft-deletes own user account
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponseStruct'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.ErrorResponseStruct'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponseStruct'
      security:
      - BearerAuth: []
      summary: Delete user (soft delete)
      tags:
      - Users
    get:
      consumes:
      - application/json
      description: Returns user details (only own profile)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponseStruct'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.ErrorResponseStruct'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponseStruct'
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - Users
    patch:
      consumes:
      - application/json
      description: Updates user profile (email, phone, password)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponseStruct'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponseStruct'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.ErrorResponseStruct'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponseStruct'
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Updates user profile (email, phone, password)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponseStruct'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponseStruct'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.ErrorResponseStruct'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponseStruct'
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - Users
  /users/email/{email}:
    get:
      consumes:
      - application/json
      description: Returns user details by email
      parameters:
      - description: User email
        in: path
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponseStruct'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponseStruct'
      security:
      - BearerAuth: []
      summary: Get user by email
      tags:
      - Users
securityDefinitions:
  BearerAuth:
    description: Type "Bearer " followed by your access token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"